Vistas de jefeSiniestro Puesto el buscador correcto y poniendo la navegacion con rutas.
