+++ pages/jefeLiquidador/siniestrosViews/nav.tsx
@@ -1,6 +1,7 @@
-import React, { useEffect } from "react";
+import React from "react";
 import styled from "styled-components";
-
+import Link from 'next/link';
+import { useRouter } from "next/router";
 import SiniestroIcon from "../../../libs/components/icons/AlertHandDrawnSymbol.component";
 import TrabajoIcon from "../../../libs/components/icons/CustomerSupport.component";
 import UserIcon from "../../../libs/components/icons/BaselineFace24px.component";
@@ -10,52 +11,28 @@ import TallerIcon from "../../../libs/components/icons/Carmechanic89096.componen
 import PresupuestoIcon from "../../../libs/components/icons/FoldedNewspaper.component";
 import MultimediaIcon from "../../../libs/components/icons/Gallery.component";
 
-interface PropsComp {
-    handleSection(sectionName: string): any;
-}
 
-
-export default ({ handleSection }: PropsComp) => {
-
-    useEffect(() => {
-        initBorder();
-        toggleStylesBtns();
-    }, []);
-
-    const toggleStylesBtns = () => {
-        const links = document.getElementsByClassName("link");
-        const removeBorder = () => {
-            for (let i = 0; i < links.length; i++) {
-                const link: any = links[i];
-                link.style.border = "none";
-            }
-        }
-        for (let i = 0; i < links.length; i++) {
-            const link: any = links[i];
-            link.addEventListener("click", () => {
-                removeBorder();
-                link.style.border = "3px solid black";
-            });
+export default () => {
+    const router = useRouter();
+    const styleActive = (href:String) => {
+        return {
+            border: router.pathname === href ? "3px solid black" : ""
         }
     }
-    const initBorder = () => {
-        const link:any = document.getElementsByClassName("link")[0];
-        link.style.border = "3px solid black";
-    }
-
-
     return (
         <LiquidadoresNav>
             {
                 links.map((item, index) =>
-                    <a onClick={() => handleSection(item.txt)} className="link" key={index}>
-                        <div className="link_icon">
-                            {item.icon}
-                        </div>
-                        <span>
-                            {item.txt}
-                        </span>
-                    </a>
+                    <Link key={index} href={item.url}>
+                        <a className="link" style={styleActive(item.url)}>
+                            <div className="link_icon">
+                                {item.icon}
+                            </div>
+                            <span>
+                                {item.txt}
+                            </span>
+                        </a>
+                    </Link>
                 )
             }
         </LiquidadoresNav>
@@ -93,41 +70,49 @@ const LiquidadoresNav = styled.div`
 
 const links = [
     {
+        url: "/jefeLiquidador/siniestrosViews/perfilLiquidador",
         icon: <UserIcon />,
         txt: "liquidador",
 
     },
     {
+        url: "",
         icon: <PolizaIcon />,
         txt: "poliza",
 
     },
     {
+        url: "/jefeLiquidador/siniestrosViews/siniestro",
         icon: <SiniestroIcon />,
         txt: "siniestro",
 
     },
     {
+        url: "",
         icon: <VehiculoIcon />,
         txt: "vehículos",
 
     },
     {
+        url: "",
         icon: <TrabajoIcon />,
         txt: "trabajos",
 
     },
     {
+        url: "/jefeLiquidador/siniestrosViews/taller",
         icon: <TallerIcon />,
         txt: "taller",
 
     },
     {
+        url: "",
         icon: <PresupuestoIcon />,
         txt: "presupuesto",
 
     },
     {
+        url: "",
         icon: <MultimediaIcon />,
         txt: "multimedia",
